Don't import defaults.less inside the core framework
As there's no LESS equivalent for SASS !defaults, importing defaults.less in the core framework will override any custom variable of the user.
It's the role of the user to import this file in his setup this way:

@import "inuit.css/defaults";
@import "vars";
@import "inuit.css/inuit"

Anyway, it's commented instead of removed so people who just download the framework can uncomment this line before using it.