LESS 1.5.0 @line-height-ratio fix
In LESS 1.5.0, @line-height-ratio is incorrectly evaluated with a 'px'
unit added to the end. This causes unexpected results mainly with its
use in base/main.less for the font property of the html selector, where
line height is determined to be 1.6px rather than 1.6 as a ratio.

Wrapping the equation in LESS's 'unit' function, without the optional
second unit parameter ensures the output is a unitless integer